account payment: fixes to canceled payments
